requirements.txt
boto3
python-dotenv
pillow
tweepy
pandas
numpy
pytz
